refactor: move deployer to utils
